Change how highlight is handled to remove it from elements when needed
Selecting rows in a reorderable list now removes the highlight from the previously focused control
